feat: allow direct governance unpause
Even if pause function is intentended to be use as a panic
button for a black swam event. It might be natural that after
it is apply, in case of a bug for example, the community shold
came up with a fix, apply it via changer, and unpause the protocol.